requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client
openai
pypdf
python-dotenv


